v1.0.0. React 16 supported. TSLint rules added to the repository. Public React typings used.
React 16 forbids to update this.state in componentWillMount.
In one of the previous versions, we moved the initial state build in this method.

This was done to avoid calling the _buildState on a partially constructed object.

Since this commit, this.state will be undefined in componentWillMount.
You could get an access to the initialState by overriding the _buildInitialState method.

We are testing following pattern on skype codebase now:

```
import { ComponentBase } from 'resub';

export abstract class InitialStateComponent<P, S> extends ComponentBase<P, S> {

    private _initialState: Readonly<S>;
    protected _buildInitialState(): Readonly<S> {
        this._initialState = super._buildInitialState();
        this.componentMounting(this._initialState);
        return this._initialState;
    }
    // tslint:disable-next-line:override-calls-super
    componentMounting(initialState: Readonly<S>): void {
        // left empty to enable super rule for this method
    }
}

```

If it will be successful we are going to add this new lifecycle method to resub and deprecate componentWillMount.